+++ packages/at_lookup/lib/src/cache/cacheable_secondary_address_finder.dart
@@ -18,9 +18,9 @@ class CacheableSecondaryAddressFinder implements SecondaryAddressFinder {
   late SecondaryUrlFinder _secondaryFinder;
 
   CacheableSecondaryAddressFinder(this._rootDomain, this._rootPort,
-      {SecondaryUrlFinder? secondaryFinder}) {
-    _secondaryFinder =
-        secondaryFinder ?? SecondaryUrlFinder(_rootDomain, _rootPort);
+      {SecondaryUrlFinder? secondaryFinder, SecureSocketConfig? socketConfig}) {
+    _secondaryFinder = secondaryFinder ??
+        SecondaryUrlFinder(_rootDomain, _rootPort, socketConfig: socketConfig);
   }
 
   bool cacheContains(String atSign) {
@@ -133,10 +133,13 @@ class SecondaryUrlFinder {
   final String _rootDomain;
   final int _rootPort;
   late final AtLookupSecureSocketFactory _socketFactory;
+  late final SecureSocketConfig _socketConfig;
 
   SecondaryUrlFinder(this._rootDomain, this._rootPort,
-      {AtLookupSecureSocketFactory? socketFactory}) {
+      {AtLookupSecureSocketFactory? socketFactory,
+      SecureSocketConfig? socketConfig}) {
     _socketFactory = socketFactory ?? AtLookupSecureSocketFactory();
+    _socketConfig = socketConfig ?? SecureSocketConfig();
   }
 
   final _logger = AtSignLogger('AtServerUrlFinder');
@@ -191,7 +194,10 @@ class SecondaryUrlFinder {
       var once = true;
 
       socket = await _socketFactory.createSocket(
-          _rootDomain, '$_rootPort', SecureSocketConfig());
+        _rootDomain,
+        '$_rootPort',
+        _socketConfig,
+      );
       _logger.finer('findAtServerUrl: connection to atDirectory established');
       // listen to the received data event stream
       socket.listen((List<int> event) async {